DRAPP-665
Performance improvements in telemetry manager.
modified code for fetching data from the telemetry and computing the shed to remove iterations.
Added an ejb single action timer for firing 30 minutes after the event is over to fetch the past 1 hour events and compute the shed again.
The reports will compute the shed immediately and after 30 minutes to ensure the data in reports are correct.
Added a reportEAO method to return the resource performance objects based on the event entityid.